+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- Added `GetPendingSalesOrderData` endpoint in `BusinessCore()` source.
+- Added `GetPendingSalesOrderData`, `GetSalesInvoiceData`, `GetSalesReturnDetailData` 
+endpoints in `BusinessCore()` source.
 
 ## [0.4.15] - 2023-05-11
 ### Added